% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpacks.R
\name{loadpacks}
\alias{loadpacks}
\title{Load R packages}
\usage{
loadpacks(pkg, ...)
}
\arguments{
\item{pkg}{R package to be loaded}
}
\value{

}
\description{
This function loads R packages easily.
}
\examples{
loadpacks(plyr, tidyverse)

}
\keyword{loadpacks}
